Improve phase2TrackerMonitorDigi: - merge header file in the plugin implementation - const correctness of input parameter - LogInfo -> LogDebug (to avoid large amount of allocations)
